% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriverWGS.R
\name{ActiveDriverWGS}
\alias{ActiveDriverWGS}
\title{ActiveDriverWGS is a driver discovery tool for simple somatic mutations in cancer whole genomes}
\usage{
ActiveDriverWGS(mutations, elements, sites = NULL, window_size = 50000,
  filter_hyper_MB = 30, recovery.dir = "", mc.cores = 1)
}
\arguments{
\item{mutations}{}

\item{elements}{}

\item{sites}{}

\item{window_size}{}

\item{filter_hyper_MB}{}

\item{recovery.dir}{}

\item{mc.cores}{}
}
\description{
ActiveDriverWGS is a driver discovery tool for simple somatic mutations in cancer whole genomes
}
\examples{
TODO Make sure all mutations are legal
TODO Format all results
}
