% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADWGS_test.R
\name{ADWGS_test}
\alias{ADWGS_test}
\title{ADWGS_test executes the statistical test for ActiveDriverWGS}
\usage{
ADWGS_test(id, gr_element_coords, gr_site_coords, gr_maf, win_size)
}
\arguments{
\item{id}{}

\item{gr_element_coords}{}

\item{gr_site_coords}{}

\item{gr_maf}{}

\item{win_size}{}
}
\description{
ADWGS_test executes the statistical test for ActiveDriverWGS
}
