% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastcancer-data.R
\docType{data}
\name{breastcancer_mutations}
\alias{breastcancer_mutations}
\title{breast cancer mutations}
\source{
\href{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl}{FTP Server}
}
\usage{
data(breastcancer_mutations)
}
\description{
breast cancer whole genome simple somatic mutations from Alexandrov et, 2013
}
\examples{
data(breastcancer_mutations)
data(cds_coords)
ActiveDriverWGS(mutations = breastcancer_mutations, elements = cds_coords)
}
\references{
Alexandrov, Ludmil B., et al. "Signatures of mutational processes in human cancer." Nature 500.7463 (2013): 415.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23945592}{PubMed})
}
\keyword{datasets}
