% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADWGS_test.R
\name{.assign_muts}
\alias{.assign_muts}
\title{.assign_muts creates a dataframe with positions, 3N context and # of mutations for 
the entire region}
\usage{
.assign_muts(gr_maf_element_plus_background, gr_background_only,
  gr_element_coords, gr_site_coords, element_bias = T, site_bias = T)
}
\arguments{
\item{gr_maf_element_plus_background}{A GRanges object containing mutations in the region
and in the background. It contains the mcols \code{c(mcols.patient, mcols.tag)}}

\item{gr_background_only}{A GRanges object containing the coordinates of the background.}

\item{gr_element_coords}{A GRanges object containing the coordinates of the region.
It contains the mcols \code{mcols} representing element ID}

\item{gr_site_coords}{A GRanges object containing the coordinates of the sites.
It contains the mcols \code{mcols} representing element ID}

\item{element_bias}{Whether or not indels overlapping the boundaries of the element
and the background should be biased towards the element}

\item{site_bias}{Whether or not indels overlapping the boundaries of the sites
should be biased towards to sites}
}
\value{
A data.frame called \code{muts_per_pos} containing the columns
pos1, tag and n_mut indicating the position of the mutation, the 3N context
and the number of mutations at that position with the same 3N context
}
\description{
.assign_muts creates a dataframe with positions, 3N context and # of mutations for 
the entire region
}
