% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADWGS_test.R
\name{.get_obs_exp}
\alias{.get_obs_exp}
\title{Calculates the number of expected mutations based}
\usage{
.get_obs_exp(hyp, select_positions, dfr, colname)
}
\arguments{
\item{hyp}{hypothesis to be tested}

\item{select_positions}{boolean column which indicates which positions are in the element of interest}

\item{dfr}{a dataframe containing the data to be tested}

\item{colname}{name of the column which indicates the count of mutations in the positions of interest}
}
\value{
a list of observed mutations (numeric), expected mutations (numeric),
observations enriched (boolean) and observations depleted (boolean)
}
\description{
Calculates the number of expected mutations based
}
\examples{
# Don't run
.get_obs_exp(h0, dfr$is_element, dfr, "n_mut")
}
