% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_muts.R
\name{format_muts}
\alias{format_muts}
\title{Title}
\usage{
format_muts(fname, cols_we_need, sample_blacklist,
  sample_name_col = "file.name", read_nrow = 1e+08,
  filter_hyper_MB = NA)
}
\arguments{
\item{fname}{}

\item{cols_we_need}{}

\item{sample_blacklist}{}

\item{sample_name_col}{}

\item{read_nrow}{}

\item{filter_hyper_MB}{}
}
\description{
Title
}
