% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADWGS_test.R
\name{make_mut_signatures}
\alias{make_mut_signatures}
\title{Makes mutational signatures}
\usage{
make_mut_signatures()
}
\value{
a dataframe with mutational signatures
}
\description{
Makes mutational signatures
}
\examples{
mut_signatures = make_mut_signatures()
}
