% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancancer-data.R
\docType{data}
\name{pancancer_mutations}
\alias{pancancer_mutations}
\title{pancancer mutations}
\source{
\href{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl}{FTP Server}
}
\usage{
data(pancancer_mutations)
}
\description{
pancancer whole genome simple somatic mutations from Alexandrov et, 2013
}
\examples{
data(pancancer_mutations)
data(cds_coords)
ActiveDriverWGS(mutations = pancancer_mutations, elements = cds_coords)
}
\references{
Alexandrov, Ludmil B., et al. "Signatures of mutational processes in human cancer." Nature 500.7463 (2013): 415.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23945592}{PubMed})
}
\keyword{datasets}
