% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_factor_binding_sites-data.R
\docType{data}
\name{transcription_factor_binding_sites}
\alias{transcription_factor_binding_sites}
\title{enhancer regions}
\format{A data frame containing the following columns: chr, start, end, id
\describe{
    \item{chr}{autosomal chromosomes as chr1 to chr22 and sex chromosomes as chrX and chrY}
    \item{start}{the start position of the site in base 0 coordinates (BED format)}
    \item{end}{the end position of the site in base 0 coordinates (BED format)}
    \item{id}{the site identifier - each site should contain only 1 segment and a unique id. If ids are duplicated, each segment of the site will be treated as an individual site. Sites can be coding or noncoding such as phosphosites of protein coding genes in genomic coordinates or transcription factor binding sites of active enhancers.}
}}
\source{
\href{https://www.biorxiv.org/content/early/2017/12/19/236802}{bioRxiv}
}
\usage{
data(transcription_factor_binding_sites)
}
\description{
enhancer regions as referenced in PCAWG
}
\examples{
data(pancancer_mutations)
data(enhancer_regions)
data(transcription_factor_binding_sites)
ActiveDriverWGS(mutations = pancancer_mutations, elements = enhancer_regions, sites = transcription_factor_binding_sites)
}
\references{
Wadi, Lina, et al. "Candidate cancer driver mutations in super-enhancers and long-range chromatin interaction networks." bioRxiv (2017): 236802.
(\href{https://www.biorxiv.org/content/early/2017/12/19/236802}{bioRxiv})
}
\keyword{datasets}
