% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastcancer-data.R
\docType{data}
\name{breastcancer_mutations}
\alias{breastcancer_mutations}
\title{breast cancer mutations}
\format{A data frame containing the following columns: chr, pos1, pos2, ref, alt, patient.
\describe{
    \item{chr}{autosomal chromosomes as chr1 to chr22 and sex chromosomes as chrX and chrY}
    \item{pos1}{the start position of the mutation in base 1 coordinates}
    \item{pos2}{the end position of the mutation in base 1 coordinates}
    \item{ref}{the reference allele as a string containing the bases A, T, C or G}
    \item{alt}{the alternate allele as a string containing the bases A, T, C or G}
    \item{patient}{the patient identifier as a string}
    \item{cc}{cancer type of the sample}
}}
\source{
\href{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl}{FTP Server}
}
\usage{
data(breastcancer_mutations)
}
\description{
breast cancer whole genome simple somatic mutations from Alexandrov et, 2013
}
\examples{
data(breastcancer_mutations)
data(coding_regions)
ActiveDriverWGS(mutations = breastcancer_mutations, elements = coding_regions)
}
\references{
Alexandrov, Ludmil B., et al. "Signatures of mutational processes in human cancer." Nature 500.7463 (2013): 415.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23945592}{PubMed})
}
\keyword{datasets}
